+++ src/toolbar.js
@@ -50,6 +50,8 @@ var defaultItems = [
 		{ type:'separator' },
 		{ type:'color', label:'Color', items:defaultColors },
 		{ type:'background', label:'Background color', items:defaultColors },
+		{ type:'separator' },
+		{ type:'link', label:'Link' }
 	]},
 
 	{ label:'Blocks', type:'group', items: [